Reorganized some files' & fixed SelectionTable bug
