+++ client/src/Components/index.js
@@ -30,6 +30,7 @@ export { default as Trash } from './UI/Trash/Trash';
 export { default as MousePointer } from './UI/Icons/MousePointer';
 export { default as ArrowLeft } from './UI/Icons/ArrowLeft';
 export { default as ArrowDown } from './UI/Icons/ArrowDown';
+export { default as ToggleGroup } from './UI/ToggleGroup/ToggleGroup';
 export { default as EditableText } from './EditableText/EditableText';
 export { default as EditText } from './EditableText/EditText';
 export { default as SelectionList } from './Form/SelectionList/SelectionList';
