+++ client/src/Containers/Monitoring/dropdownmenu.css
@@ -1,0 +1,51 @@
+:import('../../colors.css') {
+  blue: blue;
+  lightShadow: lightShadow;
+}
+
+.Container {
+  position: relative;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+}
+
+.DropdownContent {
+  display: none;
+  position: absolute;
+  top: 10px;
+  left: 10px;
+  right: 0;
+
+  z-index: 1000;
+  box-shadow: lightShadow;
+  background: blue;
+  color: white;
+  border: 1px solid #ddd;
+  width: 130px;
+  flex-flow: column;
+  transition: 0.2s;
+}
+
+.DropdownItem {
+  margin: 8px 0;
+}
+
+.Container:hover .DropdownContent {
+  display: flex;
+  animation: showDropdown 0.2s;
+}
+
+@keyframes showDropdown {
+  /* style rules */
+  0% {
+    transform: translateY(-20px);
+    opacity: 0;
+  }
+  100% {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
