+++ client/src/Containers/Monitoring/monitoringView.css
@@ -1,4 +1,4 @@
-:import('../../../colors.css') {
+:import('../../colors.css') {
   blue: blue;
   blueHover: blueHover;
   gray: backgroundGray;
