+++ client/src/Containers/index.js
@@ -34,3 +34,5 @@ export { default as SharedReplayer } from './Replayer/SharedReplayer';
 export { default as Dashboard } from './Dashboard';
 export { default as Chart } from './Stats/Chart';
 export { default as statsReducer, initialState } from './Stats/statsReducer';
+export { default as MonitoringView } from './Monitoring/MonitoringView';
+export { default as RoomMonitorView } from './Monitoring/RoomMonitorView';
